requirements.txt
Django==3.1.12
django_prometheus==2.3.1
djangorestframework==3.12.4
djongo==1.3.7
prometheus_client==0.20.0
whitenoise==6.8.2
gunicorn==23.0.0
